visuals/requirements.txt
vcdvcd==2.6.0
numpy>=1.26
pillow>=10.0
